requirements.txt
alembic==1.13.1
amqp==5.2.0
beautifulsoup4==4.12.2
billiard==4.2.0
blinker==1.7.0
bs4==0.0.1
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.4
Mako==1.3.0
MarkupSafe==2.1.3
prompt-toolkit==3.0.43
pyaes==1.6.1
pyasn1==0.5.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.23
Telethon==1.33.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
WTForms==3.1.1